+++ components/Document/Document.test.js
@@ -71,58 +71,39 @@ describe( 'Document', () => {
     expect( isOneExpanded ).toEqual( true );
   } );
 
-  it( 'renders Share PopupTrigger (Preview)', () => {
-    const shareTrigger = previewWrapper.findWhere( n => n.name() === 'PopupTrigger' && n.prop( 'toolTip' ) === 'Share document' );
+  it( 'renders Share Popup (Preview)', () => {
+    const sharePopup = previewWrapper.find( 'Popover.document-project__popover--share' );
 
-    expect( shareTrigger.exists() ).toEqual( true );
-
-    const content = mount( shareTrigger.prop( 'content' ) );
-
-    expect( content.exists() ).toEqual( true );
-    expect( content.name() ).toEqual( 'Popup' );
-    expect( content.prop( 'title' ) ).toEqual( 'Copy the link to share internally.' );
-
-    const share = content.find( 'Share' );
-
-    expect( share.exists() ).toEqual( true );
-    expect( share.props() ).toEqual( {
-      id: previewProps.item.id,
-      isPreview: previewProps.isAdminPreview,
-      language: 'en-us',
-      link: 'The direct link to the package will appear here.',
-      site: previewProps.item.site,
-      title: previewProps.item.title,
-      type: 'document',
-    } );
+    expect( sharePopup.exists() ).toEqual( true );
   } );
 
-  it( 'renders the Share PopupTrigger (Non-preview)', () => {
-    const shareTrigger = nonPreviewWrapper.findWhere( n => n.name() === 'PopupTrigger' && n.prop( 'toolTip' ) === 'Share document' );
+  it( 'renders the Share (Non-preview)', () => {
+    const sharePopup = previewWrapper.find( 'Popover.document-project__popover--share' );
 
-    expect( shareTrigger.exists() ).toEqual( true );
+    expect( sharePopup.exists() ).toEqual( true );
 
-    const content = mount( shareTrigger.prop( 'content' ) );
+    // const shareTrigger = nonPreviewWrapper.findWhere( n => n.name() === 'PopupTrigger' && n.prop( 'toolTip' ) === 'Share document' );
+    // expect( shareTrigger.exists() ).toEqual( true );
 
-    expect( content.exists() ).toEqual( true );
-    expect( content.name() ).toEqual( 'Popup' );
-    expect( content.prop( 'title' ) ).toEqual( 'Copy the link to share internally.' );
+    // const content = mount( shareTrigger.prop( 'content' ) );
+    // expect( content.exists() ).toEqual( true );
+    // expect( content.name() ).toEqual( 'Popup' );
+    // expect( content.prop( 'title' ) ).toEqual( 'Copy the link to share internally.' );
 
-    const share = content.find( 'Share' );
+    // const share = content.find( 'Share' );
+    // expect( share.exists() ).toEqual( true );
+    // expect( share.props() ).toEqual( {
+    //   id: nonpreviewProps.item.id,
+    //   isPreview: nonpreviewProps.isAdminPreview,
+    //   language: 'en-us',
+    //   site: nonpreviewProps.item.site,
+    //   title: nonpreviewProps.item.title,
+    //   type: 'document',
+    // } );
 
-    expect( share.exists() ).toEqual( true );
-    expect( share.props() ).toEqual( {
-      id: nonpreviewProps.item.id,
-      isPreview: nonpreviewProps.isAdminPreview,
-      language: 'en-us',
-      site: nonpreviewProps.item.site,
-      title: nonpreviewProps.item.title,
-      type: 'document',
-    } );
-
-    const directLink = share.find( 'ClipboardCopy' ).prop( 'copyItem' );
-    const documentLink = `http://${nonpreviewProps.item.site}/${nonpreviewProps.item.type}?id=${nonpreviewProps.item.id}&site=${nonpreviewProps.item.site}&language=en-us`;
-
-    expect( directLink ).toEqual( documentLink );
+    // const directLink = share.find( 'ClipboardCopy' ).prop( 'copyItem' );
+    // const documentLink = `http://${nonpreviewProps.item.site}/${nonpreviewProps.item.type}?id=${nonpreviewProps.item.id}&site=${nonpreviewProps.item.site}&language=en-us`;
+    // expect( directLink ).toEqual( documentLink );
   } );
 
   it( 'renders the Download element (Preview)', () => {